+++ example/webpack/modules/todocomponent/constants/storage.js
@@ -1,0 +1,3 @@
+import {constant} from 'src/app';
+
+export default constant('storage-id', 'angular-todo-mvc-id');