+++ webpack/webpack.dev.config.js
@@ -77,7 +77,7 @@ config.plugins = [
 	new webpack.optimize.OccurenceOrderPlugin(),
 	new webpack.optimize.DedupePlugin(),
 	new webpack.NoErrorsPlugin(),
-	//new webpack.HotModuleReplacementPlugin(),
+	new webpack.HotModuleReplacementPlugin(),
 	new webpack.optimize.CommonsChunkPlugin('vendors', 'vendors.js')
 ];